% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisProject.R
\name{createAnalysisProject}
\alias{createAnalysisProject}
\title{Create a project laid out for an analysis project}
\usage{
createAnalysisProject(name, dirs = c("data", "analysis", "outputs"), ...)
}
\arguments{
\item{name}{Project}

\item{dirs}{Directories to create}

\item{...}{Pass to \code{\link[=createBasicProject]{createBasicProject()}}}
}
\description{
Create a project laid out for an analysis project
}
\examples{
\dontrun{
proj<-"analysisproj"
createAnalysisProject(proj)
list.files(proj)
unlink(proj)
}
}
