% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingProject.R
\name{createTrainingProject}
\alias{createTrainingProject}
\title{Create a project laid out for producing training materials}
\usage{
createTrainingProject(name, dirs = c("data", "handouts", "slides"),
  handoutEngine = "rmarkdown", slideEngine = "rmarkdown", ...)
}
\arguments{
\item{name}{Project}

\item{dirs}{Directories to create}

\item{handoutEngine}{Package for building handouts}

\item{slideEngine}{Package for building slides}

\item{...}{Pass to \code{\link[=createBasicProject]{createBasicProject()}}}
}
\description{
Create a project laid out for producing training materials
}
\examples{
\dontrun{
proj<-"trainingproj"
createTrainingProject(proj)
list.files(proj)
unlink(proj)
}
}
