% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_health.R
\name{get_project_health}
\alias{get_project_health}
\title{Create project health report under health}
\usage{
get_project_health()
}
\description{
Create project health report under health
}
\details{
for current working directory.
Aimed at use on CI. It currently includes \code{goodpractice}
and \code{devtools::spell_check()} output.
}
