% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_init.R
\name{lib_init}
\alias{lib_init}
\title{A Library Intialization Function}
\usage{
lib_init()
}
\description{
This function allows you to import all necessary (or frequently used) R packages .
}
\examples{
init()
}
\keyword{Stores}
\keyword{all}
\keyword{packages}
\keyword{x:}
