% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib_setup.R
\name{lib_setup}
\alias{lib_setup}
\title{A Setup Function}
\usage{
lib_setup()
}
\description{
This function allows you to install all necessary (or frequently used) R packages .
}
\examples{
setup()
}
