% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loc_init.R
\name{loc_init}
\alias{loc_init}
\title{A Location Initialization Function}
\usage{
loc_init()
}
\description{
This function allows you to set your preferred working directory .
}
\examples{
init_loc()
}
