% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_here.R
\name{start_here}
\alias{start_here}
\title{Check your setup before using starters}
\usage{
start_here()
}
\description{
Check your setup before using starters
}
\details{
This function checks different aspects of your setup,
that will help automatic steps later on. Most of these aspects
are inspired from \code{usethis} setup article,
https://usethis.r-lib.org/articles/articles/usethis-setup.html.
}
