% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starters-package.R
\docType{package}
\name{starters}
\alias{starters}
\title{starters: Manage project directories for R starters}
\description{
Setting up R project directories can be a pain. starters provides shortcut
functions which can help with this by creating template file and folder
structures and setting good defaults for analysis, training, or package
development starters.
}
\details{
To learn more about starters, start with the vignettes: browseVignettes(package = "starters")
}
