% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_news_md}
\alias{use_news_md}
\title{Revised function to not try to open file}
\usage{
use_news_md(pkg = ".")
}
\arguments{
\item{pkg}{See \code{\link[devtools:use_news_md]{devtools::use_news_md()}}}
}
\description{
Revised function to not try to open file
}
