% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_package_doc}
\alias{use_package_doc}
\title{Revised function to not try to open file}
\usage{
use_package_doc(pkg = ".")
}
\arguments{
\item{pkg}{See \code{\link[devtools:use_package_doc]{devtools::use_package_doc()}}}
}
\description{
Revised function to not try to open file
}
