% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_readme_rmd}
\alias{use_readme_rmd}
\title{Revised function to not try to open file}
\usage{
use_readme_rmd(pkg = ".")
}
\arguments{
\item{pkg}{See \code{\link[devtools:use_readme_rmd]{devtools::use_readme_rmd()}}}
}
\description{
Revised function to not try to open file
}
