% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisProject.R
\name{createAnalysisProject}
\alias{createAnalysisProject}
\title{Create a project laid out for an analysis project}
\usage{
createAnalysisProject(name, folder = getwd(), travis = TRUE,
  packagedeps = "packrat", git = TRUE,
  github = gh::gh_whoami()$login, private = FALSE, protocol = "ssh",
  dirs = c("data", "analysis", "outputs"), title = NULL)
}
\arguments{
\item{name}{Project}

\item{folder}{Folder under which to create the project}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{dirs}{Directories to create}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}
}
\description{
Create a project laid out for an analysis project
}
\examples{
\dontrun{
proj<-"analysisproj"
createAnalysisProject(proj)
list.files(proj)
unlink(proj)
}
}
