% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisProject.R
\name{createAnalysisProject}
\alias{createAnalysisProject}
\alias{create_analysis_project}
\title{Create a project laid out for an analysis project}
\usage{
createAnalysisProject(
  name,
  title = NULL,
  folder = getwd(),
  initial_status = "wip",
  dirs = c("data", "analysis", "outputs"),
  packagedeps = "packrat",
  git = TRUE,
  external_setup = list(git_service = "GitHub", login = gh::gh_whoami()$login, private =
    FALSE, protocol = "ssh", ci_activation = "travis"),
  reset = TRUE,
  open = FALSE
)

create_analysis_project(
  name,
  title = NULL,
  folder = getwd(),
  initial_status = "wip",
  dirs = c("data", "analysis", "outputs"),
  packagedeps = "packrat",
  git = TRUE,
  external_setup = list(git_service = "GitHub", login = gh::gh_whoami()$login, private =
    FALSE, protocol = "ssh", ci_activation = "travis"),
  reset = TRUE,
  open = FALSE
)
}
\arguments{
\item{name}{Project}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}

\item{folder}{Folder under which to create the project}

\item{initial_status}{initial repostatus.org status for the project,
whose badge will be added to the README.}

\item{dirs}{Directories to create}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{external_setup}{How to do the partly interactive setup
of online git, CI and coverage platforms. If NULL, no setup.
\itemize{
\item \code{git_service} Only "GitHub" is supported at the moment.
\item \code{login} username or organization name to use for the git service.
If NULL, no GitHub repo is created.
\item \code{private} whether to make the created repo private
\item \code{protocol} "ssh" or "https", protocol to use for GitHub
\item \code{ci_activation} Only NULL, and "travis" are supported at the moment.
"travis" means calling \code{usethis::use_travis()} and adding the Travis
badge to the README.
}}

\item{reset}{Whether to reset the project to current project}

\item{open}{Whether to open the newly created project (in another RStudio session)}
}
\description{
Create a project laid out for an analysis project
}
\section{Functions}{
\itemize{
\item \code{create_analysis_project}: aesthetic addition for continuity purposes
}}

\examples{
\dontrun{
folder <- tempdir()
createAnalysisProject(
  name = "doggos", title = "Counting cute dogs",
  folder = folder,
  packagedeps = "none",
  git = TRUE, external_setup = NULL,
  reset = TRUE,
  dirs = c("cats", "dogs"),
  open = FALSE
)
list.files(file.path(folder, "doggos"))
unlink(file.path(folder, "doggos"))
}
}
