% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAnalysisProject.R
\name{createAnalysisProject}
\alias{createAnalysisProject}
\title{Create a project laid out for an analysis project}
\usage{
createAnalysisProject(name, folder = getwd(), travis = TRUE,
  packagedeps = "packrat", git = TRUE,
  github = gh::gh_whoami()$login, private = FALSE, protocol = "ssh",
  readme = TRUE, dirs = c("data", "analysis", "outputs"))
}
\arguments{
\item{name}{Project}

\item{folder}{Folder under which to create the project}

\item{travis}{Configure Travis-CI}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{github}{username or organization name to use for GitHub.
If NULL, no GitHub repo is created.}

\item{private}{whether to make the created GitHub repo private}

\item{protocol}{"ssh" or "https", protocol to use for GitHub}

\item{readme}{Include a README}

\item{dirs}{Directories to create}
}
\description{
Create a project laid out for an analysis project
}
\examples{
\dontrun{
proj<-"analysisproj"
createAnalysisProject(proj)
list.files(proj)
unlink(proj)
}
}
