% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageProject.R
\name{createPackageProject}
\alias{createPackageProject}
\title{Create a package the devtools way with a number of infrastructure
commands executed to save you hassle}
\usage{
createPackageProject(name, bestPractices = TRUE, coverage = c("codecov",
  "coveralls"))
}
\arguments{
\item{name}{Project / package name}

\item{bestPractices}{Run the additional \code{\link[devtools:infrastructure]{devtools::infrastructure()}} commands}

\item{coverage}{What code coverage platform to use}
}
\description{
Create a package the devtools way with a number of infrastructure
commands executed to save you hassle
}
\examples{
\dontrun{
proj<-"packageproj"
createPackageProject(proj)
list.files(proj)
unlink(proj)
}
}
