% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageProject.R
\name{createPackageProject}
\alias{createPackageProject}
\title{Create a package the usethis way with a number of infrastructure
commands executed to save you hassle}
\usage{
createPackageProject(name, folder = getwd(), bestPractices = TRUE,
  coverage = "codecov", github = gh::gh_whoami()$login,
  private = FALSE, protocol = "ssh", title = NULL)
}
\arguments{
\item{name}{Project / package name}

\item{folder}{Folder under which to create the project}

\item{bestPractices}{Run additional best practice commands}

\item{coverage}{What code coverage platform to use, "codecov" or "coveralls".}

\item{github}{username or organization name to use for GitHub.
If NULL, no GitHub repo is created.}

\item{private}{whether to make the created GitHub repo private}

\item{protocol}{"ssh" or "https", protocol to use for GitHub}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}
}
\description{
Create a package the usethis way with a number of infrastructure
commands executed to save you hassle
}
\examples{
\dontrun{
proj<-"packageproj"
createPackageProject(proj)
list.files(proj)
unlink(proj)
}
}
