% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageProject.R
\name{createPackageProject}
\alias{createPackageProject}
\alias{create_package_project}
\title{Create a package the usethis way with a number of infrastructure
commands executed to save you hassle}
\usage{
createPackageProject(name, title = NULL, folder = getwd(),
  initial_status = "wip", coverage = "codecov", git = TRUE,
  pkgdown = TRUE, external_setup = list(git_service = "GitHub", login =
  gh::gh_whoami()$login, private = FALSE, protocol = "ssh", ci_activation =
  "travis"), reset = TRUE, open = FALSE)

create_package_project(name, title = NULL, folder = getwd(),
  initial_status = "wip", coverage = "codecov", git = TRUE,
  pkgdown = TRUE, external_setup = list(git_service = "GitHub", login =
  gh::gh_whoami()$login, private = FALSE, protocol = "ssh", ci_activation =
  "travis"), reset = TRUE, open = FALSE)
}
\arguments{
\item{name}{Project / package name}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}

\item{folder}{Folder under which to create the project}

\item{initial_status}{initial repostatus.org status for the project,
whose badge will be added to the README.}

\item{coverage}{What code coverage platform to use, "codecov" or "coveralls".}

\item{git}{Configure Git}

\item{pkgdown}{Add a pkgdown config file}

\item{external_setup}{How to do the partly interactive setup
of online git, CI and coverage platforms. If NULL, no setup.
\itemize{
\item \code{git_service} Only "GitHub" is supported at the moment.
\item \code{login} username or organization name to use for the git service.
If NULL, no GitHub repo is created.
\item \code{private} whether to make the created repo private
\item \code{protocol} "ssh" or "https", protocol to use for GitHub
\item \code{ci_activation} Only NULL, and "travis" are supported at the moment.
"travis" means calling \code{usethis::use_travis()} and adding the Travis
badge to the README.
}}

\item{reset}{Whether to reset the project to current project}
}
\description{
Create a package the usethis way with a number of infrastructure
commands executed to save you hassle
}
\section{Functions}{
\itemize{
\item \code{create_package_project}: aesthetic addition for continuity purposes
}}

\examples{
\dontrun{
folder <- tempdir()
createPackageProject(
  name = "doggos", title = "Counting cute dogs",
  folder = folder,
  git = TRUE, external_setup = NULL,
  reset = TRUE,
  open = FALSE
)
list.files(file.path(folder, "doggos"))
unlink(file.path(folder, "doggos"))
}
}
