% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPackageProject.R
\name{createPackageProject}
\alias{createPackageProject}
\title{Create a package the usethis way with a number of infrastructure
commands executed to save you hassle}
\usage{
createPackageProject(name, folder = getwd(), bestPractices = TRUE,
  coverage = "codecov", private = TRUE)
}
\arguments{
\item{name}{Project / package name}

\item{folder}{Folder under which to create the project}

\item{bestPractices}{Run additional best practice commands}

\item{coverage}{What code coverage platform to use, "codecov" or "coveralls".}

\item{private}{Whether to create the GH repo as private}
}
\description{
Create a package the usethis way with a number of infrastructure
commands executed to save you hassle
}
\examples{
\dontrun{
proj<-"packageproj"
createPackageProject(proj)
list.files(proj)
unlink(proj)
}
}
