% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingProject.R
\name{createTrainingProject}
\alias{createTrainingProject}
\title{Create a training project}
\usage{
createTrainingProject(name, folder = getwd(), dirs = c("data",
  "handouts", "slides"), handoutEngine = "rmarkdown",
  slideEngine = "rmarkdown", travis = TRUE, packagedeps = "packrat",
  git = TRUE, github = gh::gh_whoami()$login, private = FALSE,
  protocol = "ssh", readme = TRUE)
}
\arguments{
\item{name}{Project}

\item{folder}{Folder under which to create the project}

\item{dirs}{Directories to create}

\item{handoutEngine}{Package for building handouts}

\item{slideEngine}{Package for building slides}

\item{travis}{Configure Travis-CI}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{github}{username or organization name to use for GitHub.
If NULL, no GitHub repo is created.}

\item{private}{whether to make the created GitHub repo private}

\item{protocol}{"ssh" or "https", protocol to use for GitHub}

\item{readme}{Include a README}
}
\description{
Create a project laid out for producing training materials
}
\examples{
\dontrun{
proj<-"trainingproj"
createTrainingProject(proj)
list.files(proj)
unlink(proj)
}
}
