% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingProject.R
\name{createTrainingProject}
\alias{createTrainingProject}
\title{Create a project laid out for producing training materials}
\usage{
createTrainingProject(name, dirs = c("data", "handouts", "slides"),
  handoutEngine = "rmarkdown", slideEngine = "rmarkdown",
  travis = TRUE, packagedeps = "packrat", git = TRUE,
  readme = TRUE)
}
\arguments{
\item{name}{Project}

\item{dirs}{Directories to create}

\item{handoutEngine}{Package for building handouts}

\item{slideEngine}{Package for building slides}

\item{travis}{Configure Travis-CI}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{readme}{Include a README}
}
\description{
Create a project laid out for producing training materials
}
\examples{
\dontrun{
proj<-"trainingproj"
createTrainingProject(proj)
list.files(proj)
unlink(proj)
}
}
