% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTrainingProject.R
\name{createTrainingProject}
\alias{createTrainingProject}
\title{Create a training project}
\usage{
createTrainingProject(name, folder = getwd(), dirs = c("data",
  "handouts", "slides"), handoutEngine = "rmarkdown",
  slideEngine = "rmarkdown", travis = TRUE, packagedeps = "packrat",
  git = TRUE, github = gh::gh_whoami()$login, private = FALSE,
  protocol = "ssh", title = NULL)
}
\arguments{
\item{name}{Project}

\item{folder}{Folder under which to create the project}

\item{dirs}{Directories to create}

\item{handoutEngine}{Package for building handouts}

\item{slideEngine}{Package for building slides}

\item{packagedeps}{Set a tool for package reproducibility}

\item{git}{Configure Git}

\item{private}{whether to make the created GitHub repo private}

\item{protocol}{"ssh" or "https", protocol to use for GitHub}

\item{title}{"What the Project Does (One Line, Title Case)"
If NULL, a random one will be generated.}
}
\description{
Create a project laid out for producing training materials
}
\examples{
\dontrun{
proj<-"trainingproj"
createTrainingProject(proj)
list.files(proj)
unlink(proj)
}
}
